requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
slack-sdk>=3.26.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.108.0 